unblock domains properly
